Fix igText build failure
Weirdly, igText is missing in the auto-generated zig cimgui bindings,
but both igTextEx and igTextExEx are present. This fixes the example
code so non-docking builds work correctly. Docking builds are failing for
a different reason [1], which I don't yet know the solution for.

Validated the commit here runs fine for both zig 0.15.2 and
0.16.0-dev.1484+d0ba6642b

[1]

```bash
.zig-cache\o\bbc92ceb91ff817d43e74aa248a5e988\cimgui_all.zig:1927:21: error: opaque types have unknown size and therefore cannot be directly embedded in structs
    BoxSelectState: ImGuiBoxSelectState = @import("std").mem.zeroes(ImGuiBoxSelectState),
                    ^~~~~~~~~~~~~~~~~~~
.zig-cache\o\bbc92ceb91ff817d43e74aa248a5e988\cimgui_all.zig:1437:42: note: opaque declared here
pub const struct_ImGuiBoxSelectState_t = opaque {};
```